+++ src/components/traits/i-data-provider/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v4.0.0-beta.?? (2024-??-??)
+
+#### :rocket: New Feature
+
+* Added ability to provide an async options to the `waitPermissionToRequest` method.
+
 ## v4.0.0-alpha.1 (2022-12-14)
 
 #### :rocket: New Feature